Pushing testing assets
I set up a testing scene with self explanatory scripts as a proof of concept for grid nodes, edges, and its traversal. It does not work 100% of the time but it is very close. I am holding off on pushing my changes to the final proj content as It is no closer to the solution